requirements.txt
asgiref==3.8.1
attrs==24.3.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
coverage==7.6.10
cryptography==44.0.0
defusedxml==0.8.0rc2
Django==5.1.4
django-environ==0.11.2
django-filter==24.3
django-oauth-toolkit==3.0.1
djangorestframework==3.15.2
drf-access-policy==1.5.0
drf-social-oauth2==3.1.0
gunicorn==23.0.0
idna==3.10
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jwcrypto==1.5.6
oauthlib==3.2.2
packaging==24.2
psycopg==3.2.3
psycopg-binary==3.2.3
pycparser==2.22
PyJWT==2.10.1
pyparsing==3.2.0
python3-openid==3.2.0
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.22.3
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
whitenoise==6.8.2
